requirements.txt
# Snowflake connector
snowflake-connector-python==3.6.0

# Environment variables
python-dotenv==1.0.0

# For RAG implementation
langchain==0.1.0
langchain-community==0.0.10
openai==1.7.1

# Vector embeddings
sentence-transformers==2.2.2

# Data processing
pandas==2.1.4
numpy==1.24.3
